Merge from receipt-visibility to main

* commit '7e95fd527b038919cf7f8306adbdcdaaffbcc8f7': (28 commits)
  Created logic to only show public recipes and the users own recipes
  Added Image to recipecontroller
  fixed double display of url and user in _recipespartial
  Added parent view model for user recipes, this can also take in more models if needed
  Added parent view model for users and recipes instead of just recipes
  Connected test data to Users, without connection recipe search does not work
  Added UserRecipes to Homecontroller, now it returns the family view
  added link to full recipe and image
  fixed food-image
  Added NotFound(); to homecontroller
  Satt inn enheter i testdata:ingrediensliste for å teste søk
  Added partial view for recipesearch(homepage)
  Added testdata for search
  Added search to home-controller with functions
  Connected Popular Recipes to model and Controller
  Added receipt visibility function
  Created models, logic and views for 'Tilbakemeldinger'
  Created view and controller for 'Mine Oppskrifter'
  Designed 'Ny Oppskrift' and extracted the profile nav-menu into a Partial View
  c4637b5d8d8b595d0c5dad45459e01e980c5b321
  ...

+++ Controllers/RecipesController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
@@ -25,7 +26,11 @@ namespace smartmat.Controllers
         // GET: Recipes
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Recipes.ToListAsync());
+            var user = await _userManager.GetUserAsync(User);
+            var recipes = _context.Recipes
+                .Where(recipe => recipe.Visibility == "Public" || recipe.ApplicationUser == user)
+                .ToList(); 
+            return View(recipes);
         }
 
         // GET: Recipes/Details/5
@@ -36,7 +41,9 @@ namespace smartmat.Controllers
                 return NotFound();
             }
 
+            var user = await _userManager.GetUserAsync(User);
             var recipe = await _context.Recipes
+                .Where(recipe => recipe.Visibility == "Public" || recipe.ApplicationUser == user)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (recipe == null)
             {
@@ -47,6 +54,7 @@ namespace smartmat.Controllers
         }
 
         // GET: Recipes/Create
+        [Authorize]
         public IActionResult Create()
         {
             return View();
@@ -57,7 +65,8 @@ namespace smartmat.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Title,Introduction,Ingredients,Instructions,Nutrients,Image")] Recipe recipe)
+        [Authorize]
+        public async Task<IActionResult> Create([Bind("Id,Title,Introduction,Ingredients,Instructions,Nutrients,Visibility,Image")] Recipe recipe)
         {
             if (ModelState.IsValid)
             {
@@ -73,17 +82,22 @@ namespace smartmat.Controllers
         }
 
         // GET: Recipes/Edit/5
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-
+            var user = await _userManager.GetUserAsync(User);
             var recipe = await _context.Recipes.FindAsync(id);
             if (recipe == null)
             {
                 return NotFound();
+            }
+            if (recipe.ApplicationUser != user)
+            {
+                return RedirectToAction(nameof(NotAllowed));
             }
             return View(recipe);
         }
@@ -93,45 +107,55 @@ namespace smartmat.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Introduction,Ingredients,Instructions,Nutrients,Image")] Recipe recipe)
+        [Authorize]
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Introduction,Ingredients,Instructions,Nutrients,Visibility,Image")] Recipe recipe)
         {
-            if (id != recipe.Id)
+            var user = await _userManager.GetUserAsync(User);
+            var recipeInDb = _context.Recipes
+                .Any(r => r.ApplicationUser == user && r.Id == id);
+            if (id != recipe.Id || !recipeInDb)
             {
-                return NotFound();
+                return RedirectToAction(nameof(NotAllowed));
             }
 
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                try
-                {
-                    _context.Update(recipe);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!RecipeExists(recipe.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return View(recipe);
             }
-            return View(recipe);
+            
+            try
+            {
+                recipe.ApplicationUser = user;
+                _context.Update(recipe);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!RecipeExists(recipe.Id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return RedirectToAction(nameof(Index));
+            
         }
 
         // GET: Recipes/Delete/5
+        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-
+            
+            var user = await _userManager.GetUserAsync(User);
             var recipe = await _context.Recipes
+                .Where(recipe => recipe.ApplicationUser == user)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (recipe == null)
             {
@@ -144,9 +168,16 @@ namespace smartmat.Controllers
         // POST: Recipes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var recipe = await _context.Recipes.FindAsync(id);
+            var user = await _userManager.GetUserAsync(User);
+            var recipe = await _context.Recipes
+                .FindAsync(id);
+            if (recipe.ApplicationUser != user)
+            {
+                return RedirectToAction(nameof(NotAllowed));
+            }
             _context.Recipes.Remove(recipe);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -155,6 +186,13 @@ namespace smartmat.Controllers
         private bool RecipeExists(int id)
         {
             return _context.Recipes.Any(e => e.Id == id);
+        }
+        
+        // Not Allowed
+        [HttpGet]
+        public IActionResult NotAllowed()
+        {
+            return View();
         }
     }
 }
